pyproject.toml
[project]
name = "hackrx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "pydantic-ai>=0.4.7",
    "graphiti-core[google-genai]>=0.3.21",
    "pymupdf>=1.26.3",
    "supabase>=2.17.0",
    "openai>=1.97.1",
    "aiofiles>=24.1.0",
    "motor>=3.7.1",
    "python-docx>=1.2.0",
]
